fix: 수익이 integer가 표현할 수 있는 범위를 넘어갈 때 생기는 에러를 수정
money에 대해서 long을 기본 타입으로 가지도록 수정

+++ src/main/java/lotto/domain/LottoStatistics.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import java.math.BigInteger;
 import java.util.*;
 
 import static lotto.domain.Rank.RANK_END_INDEX;
@@ -41,9 +42,9 @@ public class LottoStatistics {
     }
 
     private Money getTotalProfit() {
-        int totalProfit = 0;
+        long totalProfit = 0;
         for (Rank rank : statistics.keySet()) {
-            totalProfit += statistics.get(rank) * rank.getPrize();
+            totalProfit += (long) statistics.get(rank) * rank.getPrize();
         }
         return new Money(totalProfit);
     }
